switching back to 32-bit mode
It seems that the excessive memory usage was due to running in 64-bit mode, I think it was using the memory just because it was available. Earlier in development the application did not play well with the GC and would not run at all in 32-bit mode without throwing OOM. Earlier memory optimization seems to have resolved this, and it was only continuing in 64-bit mode that caused memory usage to seem so high. - Paul, July 9th, 2013